Update AdminListComponent location
