+++ packages/outside/src/components/AdminListComponent/UniversityList/index.tsx
@@ -1,8 +1,8 @@
 import { useGetUniversity } from '@bitgouel/api'
 import { NoneResult, WaitingAnimation } from '@bitgouel/common'
-import { UniversityDisplayInfo } from '../AdminDisplayInfo'
-import { UniversityItem } from '../AdminItemComponent'
-import * as S from './style'
+import { UniversityDisplayInfo } from '../../AdminDisplayInfo'
+import { UniversityItem } from '../../AdminItemComponent'
+import { AdminItemListContainer } from '../style'
 
 const UniversityList = () => {
   const { data, isLoading } = useGetUniversity()
@@ -10,22 +10,21 @@ const UniversityList = () => {
   return (
     <>
       <UniversityDisplayInfo />
-      <S.UserListContainer>
+      <AdminItemListContainer>
         {isLoading && <WaitingAnimation title={'대학 명단을'} />}
         {data?.universities.length <= 0 ? (
           <NoneResult title={'대학 명단이'} />
         ) : (
           data?.universities.map((university) => (
             <UniversityItem
-              key={university.id}
               universityId={String(university.id)}
               name={university.universityName}
-              nameWidth='54.5rem'
+              nameWidth='53.5rem'
               departments={university.departments}
             />
           ))
         )}
-      </S.UserListContainer>
+      </AdminItemListContainer>
     </>
   )
 }
