+++ packages/outside/src/components/AdminListComponent/index.ts
@@ -1,0 +1,2 @@
+export { default as NewUserList } from './NewUserList'
+export { default as WithdrawUserList } from './WithdrawUserList'
